Extract parseTime function to helper dir for universal time parsing
